Many minor fixes
Updated/fixed:
 - muelu settings
 - wall factor to match NASA specs
 - exp data format
 - plotting scripts

There is a factor of 1/4 in the calculation of the distance to the
nearest point off the wall, which then gets squared, leading to a
factor of 16 in the numerator. If we use a wall factor of 10/16, we
recover the NASA specs.

Values were messed up when I added the commas, also I added a rounding
of x/c for easier plotting.